requirements.txt
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
firecrawl-py==0.0.6
idna==3.7
markdownify==0.12.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
urllib3==2.2.1
